Move GetStackTrace() to instrumentation; leverage GetStackTrace() when retrieving all stack traces.
